07_Computer_Vision/04_EndtoEndProjects/services/detection_service/requirements.txt
# --- YOLOv12 Fork Dependencies (from sunsmarterjie) ---
torch==2.2.2
torchvision==0.17.2
# ⚠️ flash-attn must be in .whl file, added separately in Dockerfile
#flash_attn-2.7.3+cu11torch2.2cxx11abiFALSE-cp311-cp311-linux_x86_64.whl
timm==1.0.14
albumentations==2.0.4
onnx==1.14.0
onnxruntime==1.15.1
pycocotools==2.0.7
PyYAML==6.0.1
scipy==1.13.0
onnxslim==0.1.31
onnxruntime-gpu==1.18.0
gradio==4.44.1
opencv-python==4.9.0.80
psutil==5.9.8
py-cpuinfo==9.0.0
huggingface-hub==0.23.2
safetensors==0.4.3
numpy==1.26.4
supervision==0.22.0

# --- Ultralytics from GitHub fork ---
git+https://github.com/sunsmarterjie/yolov12.git@3a336a4adf3683d280e1a50d03fa24bbe7f24a5b


# Your service dependencies
pika
python-dotenv
psycopg2-binary
shapely
requests



